+++ docs/source/changelog.rst
@@ -47,6 +47,7 @@ Fixed
 - (Breaking) Have `x_shift` represent horizontal shifts and `y_shift` vertical shifts in :class:`deepinv.transform.Shift` (:gh:`1236` by `Sarra Amiri`_)
 - Force trainer non_blocking_transfers=False on MPS and CPU (:gh:`1311` by `Andrew Wang`_)
 - Fix :class:`deepinv.physics.PET` incorrect device attribution of attenuation and background on update and incorrect handling of batched attenuation 
+- Stop failed backtracking checks from consuming global iterations in :class:`deepinv.optim.FixedPoint`, and report the backtracking limit instead of an undefined attribute when it is reached (:gh:`PRNUM` by `Mohammad Sadegh Salehi`_)
 
 
 v0.4.1
@@ -677,6 +678,7 @@ Changed
 .. _Chaithya G R: https://github.com/chaithyagr
 .. _Alexander Skorikov: https://github.com/askorikov
 .. _Thibaut Modrzyk: https://github.com/Tmodrzyk
+.. _Mohammad Sadegh Salehi: https://github.com/MohammadSadeghSalehi
 .. _Avithal Lautman: https://github.com/avithal
 .. _Thomas Boulanger: https://github.com/LeRatonLaveurSolitaire
 .. _Tiberiu Sabau: https://github.com/tibisabau